Design Pattern
1. State Pattern
2. EncoderHandler, HttpErrorMessage.java, HttpMessage.java, OutPacketMessage.java, XHROptionsMessage.java, XHRPostMessage.java
3.
In write(ChannelHandlerContext ctx, Object msg, ChannelPromise promise) method, it works different by instance of msg. And other method in EncoderHandler.java is dependent of each instance of msg.
So I make "State" abstract class and "Concrete State" classes.
