+++ src/main/java/com/corundumstudio/socketio/messages/HttpErrorMessage.java
@@ -15,6 +15,15 @@
  */
 package com.corundumstudio.socketio.messages;
 
+import com.corundumstudio.socketio.Transport;
+import com.corundumstudio.socketio.protocol.PacketEncoder;
+import io.netty.buffer.ByteBuf;
+import io.netty.buffer.ByteBufOutputStream;
+import io.netty.channel.ChannelHandlerContext;
+import io.netty.channel.ChannelPromise;
+import io.netty.handler.codec.http.HttpResponseStatus;
+
+import java.io.IOException;
 import java.util.Map;
 
 public class HttpErrorMessage extends HttpMessage {
@@ -29,5 +38,13 @@ public class HttpErrorMessage extends HttpMessage {
     public Map<String, Object> getData() {
         return data;
     }
-    
+
+    public void messageWrite(Object message, ChannelHandlerContext ctx, ChannelPromise promise, PacketEncoder encoder) throws IOException {
+        HttpErrorMessage errorMsg = (HttpErrorMessage) message;
+        final ByteBuf encBuf = encoder.allocateBuffer(ctx.alloc());
+        ByteBufOutputStream out = new ByteBufOutputStream(encBuf);
+        encoder.getJsonSupport().writeValue(out, errorMsg.getData());
+
+        sendMessage(errorMsg, ctx.channel(), encBuf, "application/json", promise, HttpResponseStatus.BAD_REQUEST);
+    }
 }
